# swagger.yml
swagger: "2.0"
info:
  description: "Small project management API built for an interview tech challenge."
  version: "1.0.0"
  title: "airtouch challenge"
  contact:
    email: "lazugeorgem@gmail.com"
host: "localhost:3000"
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
basePath: "/"
tags:
  - name: "Authentication"
    description: "Authentication endpoints"
  - name: "User"
    description: "User endpoints"
  - name: "Project"
    description: "Project endpoints"
  - name: "Role"
    description: "Role endpoints"
paths:
  /register:
    post:
      tags:
        - "Authentication"
      summary: "Create an account"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User object to be registered on the board."
          required: true
          schema:
            $ref: "#/definitions/RegistrationRequest"
      responses:
        201:
          description: "Registered successfully"
          headers:
            Location:
              type: string
              description: "The URL of the newly created user"
          schema:
            $ref: "#/definitions/AuthenticationResponse"
        400:
          description: "Your request body was malformed"
        403:
          description: "Username already taken"
        500:
          description: "An unknown error occured while processing your request"
  /login:
    post:
      tags:
        - "Authentication"
      summary: "Login into your account"
      description: "Authenticate to your account using your username and password. You will receive a JWT that you can use in your future requests to prove your identity."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Your username and password"
          required: true
          schema:
            $ref: "#/definitions/LoginRequest"
      responses:
        200:
          description: "Login successful, JWT received"
          schema:
            $ref: "#/definitions/LoginResponse"
        400:
          description: "Your request body was malformed"
        500:
          description: "An unknown error occured while processing your request"
  /users/me:
    get:
      tags:
        - "User"
      summary: "Retrieves data about you"
      description: "Get all profile data about you."
      produces:
        - "application/json"
      responses:
        200:
          description: "Request successful, User received"
          schema:
            $ref: "#/definitions/User"
        401:
          description: "You do not have the required privileges to view this resource"
        400:
          description: "Your request body was malformed"
        500:
          description: "An unknown error occured while processing your request"
      security:
        - Bearer: []
  /users/{id}:
    get:
      tags:
        - "User"
      summary: "[Admin] Retrieves data about a user"
      description: "Get all profile data about a user."
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "The ID of the user you want to see info about"
          required: true
          type: "string"
          format: "ObjectID"
      responses:
        200:
          description: "Request successful, User received"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Your request body was malformed"
        401:
          description: "You do not have the required privileges to view this resource"
        404:
          description: "User does not exist"
        500:
          description: "An unknown error occured while processing your request"
      security:
        - Bearer: []
  /users:
    get:
      tags:
        - "User"
      summary: "[Admin] Retrieves all registered users"
      description: "Get profile data about all registered user."
      produces:
        - "application/json"
      responses:
        200:
          description: "Request successful, User received"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Your request body was malformed"
        401:
          description: "You do not have the required privileges to view this resource"
        404:
          description: "User does not exist"
        500:
          description: "An unknown error occured while processing your request"
      security:
        - Bearer: []
  /projects:
    get:
      tags:
        - "Project"
      summary: "Retrieves all existing projects"
      produces:
        - "application/json"
      responses:
        200:
          description: "Request successful, projects received"
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
        400:
          description: "Your request body was malformed"
        404:
          description: "No projects exist"
        500:
          description: "An unknown error occured while processing your request"
    post:
      tags:
        - "Project"
      summary: "[Admin] Create a new project"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User object to be registered on the board."
          required: true
          schema:
            $ref: "#/definitions/Project"
      responses:
        201:
          description: "Request successful, project created"
        401:
          description: "You do not have the required privileges to view this resource"
        400:
          description: "Your request body was malformed"
        500:
          description: "An unknown error occured while processing your request"
      security:
        - Bearer: []
  /roles:
    post:
      tags:
        - "Role"
      summary: "[Admin] Create a new authorization role"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Role object"
          required: true
          schema:
            $ref: "#/definitions/Role"
      responses:
        201:
          description: "Request successful, project created"
          headers:
            Location:
              type: string
              description: "The URL of the newly created role"
        400:
          description: "Your request body was malformed"
        401:
          description: "You do not have the required privileges to view this resource"
        500:
          description: "An unknown error occured while processing your request"
      security:
        - Bearer: []
  /roles/{id}:
    get:
      tags:
        - "Role"
      summary: "[Admin] Retrieves an authorization role"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "The ID of the user you want to see info about"
          required: true
          type: "string"
          format: "ObjectID"
      responses:
        200:
          description: "Request successful, role updated"
          schema:
            $ref: "#/definitions/Role"
        400:
          description: "Your request body was malformed"
        401:
          description: "You do not have the required privileges to view this resource"
        500:
          description: "An unknown error occured while processing your request"
      security:
        - Bearer: []
    put:
      tags:
        - "Role"
      summary: "[Admin] Modify an existing authorization role"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "The ID of the user for which you want to modify role."
          required: true
          type: "string"
          format: "ObjectID"
        - in: "body"
          name: "body"
          description: "Role object"
          required: true
          schema:
            $ref: "#/definitions/Role"
      responses:
        200:
          description: "Request successful, role updated"
          schema:
            $ref: "#/definitions/Role"
        400:
          description: "Your request body was malformed"
        401:
          description: "You do not have the required privileges to view this resource"
        500:
          description: "An unknown error occured while processing your request"
      security:
        - Bearer: []
definitions:
  User:
    type: "object"
    properties:
      _id:
        type: string
      username:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
  Project:
    type: "object"
    properties:
      _id:
        type: string
      name:
        type: string
      company:
        type: string
  Role:
    type: "object"
    properties:
      _id:
        type: string
      userId:
        type: string
      role:
        type: number
        enum: [0, 1]
  LoginRequest:
    type: "object"
    properties:
      username:
        type: string
      password:
        type: string
  AuthenticationResponse:
    type: "object"
    properties:
      token:
        type: string
  RegistrationRequest:
    type: "object"
    properties:
      username:
        type: string
      password:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string